Moving mapping functions to seperate file, mapCourseSummary now complete
